Add Interface type, distinct from Module
